requirements.txt
pyrogram==2.0.106
python-decouple==3.8
pytz==2024.1